fix(ui): Improve UI around remove txn button
